Timing attack implemented
Mostly done: it does spit out a result however due to rudimentary analysis we are very unlikely to get the right result. For now it is more of an improvement on greedy rather than a full timing attack.
It is missing a connection to the english word checker and the GUI, however the method can be run and freely tested in and of itself